colexec: minor clean up
This removes a couple of files by putting the non-templated code
into the template file.

Release justification: Category 1: Non-production code changes.

Release note: None
