--- pkg/sql/colexec/row_number.go
@@ -1,44 +1,0 @@
-// Copyright 2019 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-package colexec
-
-// TODO(yuzefovich): add benchmarks.
-
-// NewRowNumberOperator creates a new exec.Operator that computes window
-// function ROW_NUMBER. outputColIdx specifies in which exec.Vec the operator
-// should put its output (if there is no such column, a new column is
-// appended).
-func NewRowNumberOperator(input Operator, outputColIdx int, partitionColIdx int) Operator {
-	base := rowNumberBase{
-		OneInputNode:    NewOneInputNode(input),
-		outputColIdx:    outputColIdx,
-		partitionColIdx: partitionColIdx,
-	}
-	if partitionColIdx == -1 {
-		return &rowNumberNoPartitionOp{base}
-	}
-	return &rowNumberWithPartitionOp{base}
-}
-
-// rowNumberBase extracts common fields and common initialization of two
-// variations of row number operators. Note that it is not an operator itself
-// and should not be used directly.
-type rowNumberBase struct {
-	OneInputNode
-	outputColIdx    int
-	partitionColIdx int
-
-	rowNumber int64
-}
-
-func (r *rowNumberBase) Init() {
-	r.Input().Init()
-}
